+++ prisma/migrations/20250903204926_add_user_creation_and_position/migration.sql
@@ -1,0 +1,12 @@
+-- CreateTable
+CREATE TABLE "Position" (
+    "id" TEXT NOT NULL,
+    "name" TEXT NOT NULL,
+    "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    "updatedAt" TIMESTAMP(3) NOT NULL,
+
+    CONSTRAINT "Position_pkey" PRIMARY KEY ("id")
+);
+
+-- CreateIndex
+CREATE UNIQUE INDEX "Position_name_key" ON "Position"("name");